+++ src/common/greptimedb-telemetry/src/lib.rs
@@ -18,7 +18,7 @@ use std::path::{Path, PathBuf};
 use std::time::Duration;
 
 use common_runtime::error::{Error, Result};
-use common_runtime::{BoxedTaskFunction, RepeatedTask, Runtime, TaskFunction};
+use common_runtime::{BoxedTaskFunction, RepeatedTask, TaskFunction};
 use common_telemetry::{debug, info};
 use reqwest::{Client, Response};
 use serde::{Deserialize, Serialize};
@@ -49,11 +49,11 @@ impl GreptimeDBTelemetryTask {
         GreptimeDBTelemetryTask::Disable
     }
 
-    pub fn start(&self, runtime: Runtime) -> Result<()> {
+    pub fn start(&self) -> Result<()> {
         match self {
             GreptimeDBTelemetryTask::Enable(task) => {
                 print_anonymous_usage_data_disclaimer();
-                task.start(runtime)
+                task.start(common_runtime::bg_runtime())
             }
             GreptimeDBTelemetryTask::Disable => Ok(()),
         }
