+++ src/datanode/src/heartbeat.rs
@@ -16,7 +16,8 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
 
-use api::v1::meta::{HeartbeatRequest, Peer, RegionStat};
+use api::v1::meta::{HeartbeatRequest, Peer, RegionStat, Role};
+use common_grpc::channel_manager::{ChannelConfig, ChannelManager};
 use common_meta::heartbeat::handler::parse_mailbox_message::ParseMailboxMessageHandler;
 use common_meta::heartbeat::handler::{
     HandlerGroupExecutor, HeartbeatResponseHandlerContext, HeartbeatResponseHandlerExecutorRef,
@@ -24,7 +25,8 @@ use common_meta::heartbeat::handler::{
 use common_meta::heartbeat::mailbox::{HeartbeatMailbox, MailboxRef};
 use common_meta::heartbeat::utils::outgoing_message_to_mailbox_message;
 use common_telemetry::{debug, error, info, trace, warn};
-use meta_client::client::{HeartbeatSender, MetaClient};
+use meta_client::client::{HeartbeatSender, MetaClient, MetaClientBuilder};
+use meta_client::MetaClientOptions;
 use snafu::{OptionExt, ResultExt};
 use tokio::sync::mpsc;
 use tokio::time::Instant;
@@ -35,7 +37,6 @@ use crate::datanode::DatanodeOptions;
 use crate::error::{
     self, MetaClientInitSnafu, MissingMetasrvOptsSnafu, MissingNodeIdSnafu, Result,
 };
-use crate::instance::new_metasrv_client;
 use crate::region_server::RegionServer;
 
 pub(crate) mod handler;
@@ -50,7 +51,7 @@ pub struct HeartbeatTask {
     region_server: RegionServer,
     interval: u64,
     resp_handler_executor: HeartbeatResponseHandlerExecutorRef,
-    region_alive_keeper: Arc<RegionAliveKeeper>,
+    region_alive_keepers: Arc<RegionAliveKeeper>,
 }
 
 impl Drop for HeartbeatTask {
@@ -76,14 +77,14 @@ impl HeartbeatTask {
 
         let region_server = region_server.unwrap();
 
-        let region_alive_keeper = Arc::new(RegionAliveKeeper::new(
+        let region_alive_keepers = Arc::new(RegionAliveKeeper::new(
             region_server.clone(),
             opts.heartbeat.interval_millis,
         ));
         let resp_handler_executor = Arc::new(HandlerGroupExecutor::new(vec![
             Arc::new(ParseMailboxMessageHandler),
             Arc::new(RegionHeartbeatResponseHandler::new(region_server.clone())),
-            region_alive_keeper.clone(),
+            region_alive_keepers.clone(),
         ]));
 
         Ok(Self {
@@ -97,7 +98,7 @@ impl HeartbeatTask {
             region_server,
             interval: opts.heartbeat.interval_millis,
             resp_handler_executor,
-            region_alive_keeper,
+            region_alive_keepers,
         })
     }
 
@@ -162,7 +163,7 @@ impl HeartbeatTask {
         let addr = resolve_addr(&self.server_addr, &self.server_hostname);
         info!("Starting heartbeat to Metasrv with interval {interval}. My node id is {node_id}, address is {addr}.");
 
-        self.region_alive_keeper.start().await;
+        self.region_alive_keepers.start().await;
 
         let meta_client = self.meta_client.clone();
         let region_server_clone = self.region_server.clone();
@@ -180,7 +181,7 @@ impl HeartbeatTask {
         )
         .await?;
 
-        let epoch = self.region_alive_keeper.epoch();
+        let epoch = self.region_alive_keepers.epoch();
         common_runtime::spawn_bg(async move {
             let sleep = tokio::time::sleep(Duration::from_millis(0));
             tokio::pin!(sleep);
@@ -304,6 +305,39 @@ fn resolve_addr(bind_addr: &str, hostname_addr: &Option<String>) -> String {
         }
         None => bind_addr.to_owned(),
     }
+}
+
+/// Create metasrv client instance and spawn heartbeat loop.
+pub async fn new_metasrv_client(
+    node_id: u64,
+    meta_config: &MetaClientOptions,
+) -> Result<MetaClient> {
+    let cluster_id = 0; // TODO(hl): read from config
+    let member_id = node_id;
+
+    let config = ChannelConfig::new()
+        .timeout(Duration::from_millis(meta_config.timeout_millis))
+        .connect_timeout(Duration::from_millis(meta_config.connect_timeout_millis))
+        .tcp_nodelay(meta_config.tcp_nodelay);
+    let channel_manager = ChannelManager::with_config(config);
+
+    let mut meta_client = MetaClientBuilder::new(cluster_id, member_id, Role::Datanode)
+        .enable_heartbeat()
+        .enable_router()
+        .enable_store()
+        .channel_manager(channel_manager)
+        .build();
+    meta_client
+        .start(&meta_config.metasrv_addrs)
+        .await
+        .context(MetaClientInitSnafu)?;
+
+    // required only when the heartbeat_client is enabled
+    meta_client
+        .ask_leader()
+        .await
+        .context(MetaClientInitSnafu)?;
+    Ok(meta_client)
 }
 
 #[cfg(test)]
