+++ src/datanode/src/instance.rs
@@ -339,39 +339,6 @@ fn create_compaction_scheduler<S: LogStore>(opts: &DatanodeOptions) -> Compactio
     Arc::new(scheduler)
 }
 
-/// Create metasrv client instance and spawn heartbeat loop.
-pub async fn new_metasrv_client(
-    node_id: u64,
-    meta_config: &MetaClientOptions,
-) -> Result<MetaClient> {
-    let cluster_id = 0; // TODO(hl): read from config
-    let member_id = node_id;
-
-    let config = ChannelConfig::new()
-        .timeout(Duration::from_millis(meta_config.timeout_millis))
-        .connect_timeout(Duration::from_millis(meta_config.connect_timeout_millis))
-        .tcp_nodelay(meta_config.tcp_nodelay);
-    let channel_manager = ChannelManager::with_config(config);
-
-    let mut meta_client = MetaClientBuilder::new(cluster_id, member_id, Role::Datanode)
-        .enable_heartbeat()
-        .enable_router()
-        .enable_store()
-        .channel_manager(channel_manager)
-        .build();
-    meta_client
-        .start(&meta_config.metasrv_addrs)
-        .await
-        .context(MetaClientInitSnafu)?;
-
-    // required only when the heartbeat_client is enabled
-    meta_client
-        .ask_leader()
-        .await
-        .context(MetaClientInitSnafu)?;
-    Ok(meta_client)
-}
-
 pub(crate) async fn create_log_store(
     data_home: &str,
     wal_config: WalConfig,
